ile: /home/lozinski/code/cromlech/pl2/src/zopache/.git/COMMIT_EDITMSG  Modified
So backup was not restoring.
It said that some odd class was not included.
And it was a 1.5 Meg backup.  All via the parent pointer.
Setting parent pointer to null was not enough
becasue it was not using the most recent data.
So I had to delete the parent ponter, then do a  transaction.commit before back$
try except clause, to restore the parent pointer and commit the transaction in $