This is just prior to a major docker release.
I  got rid of /root/root.  It should just be /root.

Cut and Paste now displays the paste button immediately.  Previously, it required a refresh.

Right now politicians can be national or local.  There is now an option for state-wide.

Now two versions of json.  One for the discord bots, one for my legacy Single Page app.

I added a field to each page, public or private.  That allows for password
protected wiki pages.  If a page is public adding a child is public, and vice versa. Move a private page to be the child of a public page, and the private page stays private.