Moved Toot creation code to mastodon/toot.py
Now we have importedAccount, and Account, the later is the one that created the toot.
If a toot is by a remote account which does not exist, the account is created.

We now have new articles, and old articles.  Old articles just need the image. New Articles may not even get created if the URL does not work.

Toots have an update value method, which only updates if it has changed.