User Handles are now the long Mastodon handle including domain name.  This fixes a bug when two users from two different servers have the same handle. handle = "@" + email
Pug is the default Nodel.js templating language.  GreenMaps, although it is written in Python  supports it. In node, Pug templates are files.  In GreenMaps, they are objects, and can have children.  This makes it much much easier to manage complex forms, like the ones for the menu bar, by breaking the form up into smaller parts.

The new menu bar will be  View Node Tree Branch …
To enable that one needs a method to determine if the tree menu bar should be displayed.  So a new method was implemented.  hasrootPermission()

The manage form now has “click to edit/manage/view”

Removed not functioning demo buttons from some forms.