Now using /edit as the default edit command, not ckedit.
There is a view /outline to get the outline of a state from the
national map database.

JSONMarkdown has a variable _html, where the compiled html is stored.

There is a mail host view.

More control on breadCrumbsCore.  Used in search results.
Region has a webClass.

There is a candidate statement.

The not-found view, now shows the URL which was not found.